Rename `adapter-s4r` to `adapter-amqp` since this is a generic AMQP-0-9-1 implementation, which includes Pulsar+S4R plugin.
