+++ adapter-amqp/src/main/resources/conf/amqp_config.properties
