+++ adapter-amqp/src/main/resources/scenarios/amqp_msg_receiver.yaml
