+++ adapter-amqp/src/main/resources/scenarios/amqp_msg_sender.yaml
