+++ adapter-amqp/src/main/resources/scenarios/csv/binding_keys.csv
