+++ adapter-amqp/src/main/resources/scenarios/csv/routing_keys.csv
