--- adapter-s4r/pom.xml
@@ -1,73 +1,0 @@
-<!--
-  ~ Copyright (c) 2022-2023 nosqlbench
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>adapter-s4r</artifactId>
-    <packaging>jar</packaging>
-
-    <parent>
-        <artifactId>mvn-defaults</artifactId>
-        <groupId>io.nosqlbench</groupId>
-        <version>${revision}</version>
-        <relativePath>../mvn-defaults</relativePath>
-    </parent>
-
-    <name>${project.artifactId}</name>
-    <description>
-        A AMQP 0.91 driver for nosqlbench. This provides the ability to inject synthetic data
-        into an AMQP-0.91 (e.g. RabbitMQ) or an AMQP-0.91-compatible (e.g. Pulsar with S4R) system.
-    </description>
-
-    <properties>
-        <amqp.version>5.17.0</amqp.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.nosqlbench</groupId>
-            <artifactId>adapters-api</artifactId>
-            <version>${revision}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.rabbitmq</groupId>
-            <artifactId>amqp-client</artifactId>
-            <version>${amqp.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>commons-beanutils</groupId>
-            <artifactId>commons-beanutils</artifactId>
-            <version>1.9.4</version>
-        </dependency>
-
-        <!-- https://mvnrepository.com/artifact/org.apache.commons/commons-configuration2 -->
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-configuration2</artifactId>
-            <version>2.9.0</version>
-        </dependency>
-
-        <dependency>
-            <groupId>commons-io</groupId>
-            <artifactId>commons-io</artifactId>
-            <version>2.13.0</version>
-        </dependency>
-    </dependencies>
-
-</project>
