--- adapter-s4r/src/main/java/io/nosqlbench/adapter/s4r/S4RDriverAdapter.java
@@ -1,52 +1,0 @@
-/*
- * Copyright (c) 2022 nosqlbench
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.nosqlbench.adapter.s4r;
-
-import io.nosqlbench.adapter.s4r.ops.S4RTimeTrackOp;
-import io.nosqlbench.api.config.standard.NBConfigModel;
-import io.nosqlbench.api.config.standard.NBConfiguration;
-import io.nosqlbench.engine.api.activityimpl.OpMapper;
-import io.nosqlbench.engine.api.activityimpl.uniform.BaseDriverAdapter;
-import io.nosqlbench.engine.api.activityimpl.uniform.DriverAdapter;
-import io.nosqlbench.engine.api.activityimpl.uniform.DriverSpaceCache;
-import io.nosqlbench.nb.annotations.Service;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import java.util.function.Function;
-
-@Service(value = DriverAdapter.class, selector = "s4r")
-public class S4RDriverAdapter extends BaseDriverAdapter<S4RTimeTrackOp, S4RSpace> {
-    private final static Logger logger = LogManager.getLogger(S4RDriverAdapter.class);
-
-    @Override
-    public OpMapper<S4RTimeTrackOp> getOpMapper() {
-        DriverSpaceCache<? extends S4RSpace> spaceCache = getSpaceCache();
-        NBConfiguration adapterConfig = getConfiguration();
-        return new S4ROpMapper(this, adapterConfig, spaceCache);
-    }
-
-    @Override
-    public Function<String, ? extends S4RSpace> getSpaceInitializer(NBConfiguration cfg) {
-        return (s) -> new S4RSpace(s, cfg);
-    }
-
-    @Override
-    public NBConfigModel getConfigModel() {
-        return super.getConfigModel().add(S4RSpace.getConfigModel());
-    }
-}
