--- adapter-s4r/src/main/java/io/nosqlbench/adapter/s4r/exception/S4RAdapterInvalidParamException.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2022 nosqlbench
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.nosqlbench.adapter.s4r.exception;
-
-public class S4RAdapterInvalidParamException extends RuntimeException {
-
-    public S4RAdapterInvalidParamException(String paramName, String errDesc) {
-        super("Invalid setting for parameter (" + paramName + "): " + errDesc);
-    }
-
-    public S4RAdapterInvalidParamException(String fullErrDesc) {
-        super(fullErrDesc);
-    }
-}
